[bfd] Implement static library logic
Implement logic to finalize (apply and discared) relocations for
forced local symbols within the built bundle. The rest of the
relocations stay as-is.

Support for version scripts will be added in the next commit,
later followed by wrapping this created BFD with a .a archive so
to align with the static-archvie formatting of static libraries.

Signed-off-by: Eyal Itkin <eyal.itkin@gmail.com>