[RFC][ld] Introduce static bundle object support
Note: This patch is a draft for review.

Introduce a new "Static Bundle Object" which is a
relocatable object that supports symbol visibility
and finalization of local relocations. This format
is aimed to be incorporated within static archives
so to replace the current static archives of plain
object code files that are used to form static
libraries.

The full discussion about the format is available in the
generic-abi group that is responsible for the ELF standard:
https://groups.google.com/g/generic-abi/c/sT25-xfX9yc

Some questions I had about the process itself:
1. The patch is architecture dependent given that
relocations support is found in both bfd/elflink.c AND
in some arch-specific files (such as bfd/elf64-x86-64.c).
Hence, what is the standard way for marking some archs as
not yet supported? And what is the recommended process for
adding said support (especially given need to test it)?

Signed-off-by: Eyal Itkin <eyal.itkin@gmail.com>