[ld] Introduce static bundle object support
Introduce a new "Static Bundle Object" which is a
relocatable object that supports symbol visibility
and finalization of local relocations. This format
is aimed to be incorporated within static archives
so to replace the current static archives of plain
object code files that are used to form static
libraries.

The full discussion about the format is available in the
generic-abi group that is responsible for the ELF standard:
https://groups.google.com/g/generic-abi/c/sT25-xfX9yc

Signed-off-by: Eyal Itkin <eyal.itkin@gmail.com>