[ld] Allow symbol visibility support for ld -r
Introduce the "--localize-hidden" CLI flag, which will
be supported only for relocatables (ld -r).

When provided, hidden symbols will be converted from
STB_GLOBAL to STB_LOCAL when the relocatable is being
built. This change aligns the ET_REL creation to support
the same visibility semantics as that of the shared
library (ET_DYN).

This added support will enable users to use the
"ld -r --localize-hidden" outputs, wrapped inside
static (.a) archives, as an alternative format for the
existing static libraries.

The full discussion about the format is available in
the generic-abi group that is responsible for the ELF
standard. Specifically, the summarized architecture can
be found here:
https://groups.google.com/g/generic-abi/c/sT25-xfX9yc/m/J9SIthF4BgAJ

Signed-off-by: Eyal Itkin <eyal.itkin@gmail.com>