[ld] Allow relocation finalization for ld -r
Introduce the "--finalize-locals" CLI flag, which will
be supported only for relocatables (ld -r).

When provided, relocations that are associated with
local symbols, and that support finalization, will
undergo a finalization step. This change aligns the
ET_REL creation to support the same finalization
semantics as that of the shared library (ET_DYN). In
essence, if the symbols are local, and support finalize,
there is no need to keep these relocations in the binary.

This added support will enable users to properly strip
the output of "ld -r --localize-hidden --finalize-locals",
as the symbols will no longer be tied to unneeded relocations.

The full discussion about the format is available in
the generic-abi group that is responsible for the ELF
standard. Specifically, the summarized architecture can
be found here:
https://groups.google.com/g/generic-abi/c/sT25-xfX9yc/m/J9SIthF4BgAJ

Signed-off-by: Eyal Itkin <eyal.itkin@gmail.com>